# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Imóveis",
        "description": "API para gerenciamento de imóveis e contatos",
        "contact": {
            "name": "Suporte",
            "email": "suporte@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/newsletters": {
            "post": {
                "tags": [
                    "Newsletter"
                ],
                "summary": "Criar nova inscrição na newsletter",
                "description": "Endpoint para criar uma nova inscrição na newsletter do sistema",
                "operationId": "createNewsletter",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "terms_accept",
                                    "recaptchaToken"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 300,
                                        "example": "joao.silva@exemplo.com"
                                    },
                                    "terms_accept": {
                                        "description": "Confirmação de aceite dos termos de uso e política de privacidade",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "recaptchaToken": {
                                        "description": "Token do Google reCAPTCHA v3 para validação de segurança",
                                        "type": "string",
                                        "example": "03AFcWeA5X..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscrição criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "newsletters"
                                                    ]
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "attributes": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "João Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "joao.silva@exemplo.com"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Inscrição realizada com sucesso!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação do reCAPTCHA",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Falha ao verificar o recaptcha."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Como podemos te chamar"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Precisamos do seu e-mail para te manter informado"
                                                    }
                                                },
                                                "terms_accept": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "É necessário aceitar os termos de uso e política de privacidade."
                                                    }
                                                },
                                                "recaptchaToken": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Token de verificação é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}